Move Coordinator related operations to its own module
