+++ eth-test-helpers/src/contracts/index.ts
@@ -1,2 +1,4 @@
+import * as coordinator from './coordinator'
 import * as oracle from './oracle'
-export { oracle }
+
+export { oracle, coordinator }
